+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+- Add Svelte syntax highlighting for code blocks (thanks to Axios team for the tip here!)
 - Adds example writable/derived store setup for dynamic information display
 
 ## v0.7.1